+++ clang/docs/ReleaseNotes.md
@@ -825,6 +825,9 @@ latest release, please see the [Clang Web Site](https://clang.llvm.org) or the
 - Fixed a crash when parsing invalid friend declaration with storage-class specifier. (#GH186569)
 - Fixed a missing vtable for `dynamic_cast<FinalClass *>(this)` in a function template. (#GH198511)
 - Fixed an assertion failure during init-list checking of an array whose element type is an incomplete class. (#GH140685)
+- Fixed a crash when instantiating a member variable template partial specialization
+  whose primary template was left invalid after deserializing the enclosing class
+  template from a precompiled header or module built with errors. (#GH202956)
 
 #### Bug Fixes to AST Handling